@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=2.0.0"
 
 dependencies:
@@ -12,9 +12,10 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_form_builder: ^5.0.0-alpha.1
+  plugin_platform_interface: ^2.0.0
+  flutter_form_builder: ^6.0.0-nullsafety.1
   flutter_hooks: ^0.16.0
-  flutter_portal: ^0.3.0
+  flutter_portal: ^0.4.0-nullsafety.0
   cupertino_icons: ^1.0.2
   dio: ^4.0.0-prev3
   intl: ^0.17.0
@@ -28,7 +29,7 @@ dependencies:
   path_provider: ^2.0.1
   path: ^1.8.0
   equatable: ^2.0.0
-  dartz: ^0.9.2
+  dartz: ^0.10.0-nullsafety.1
 
 dev_dependencies:
   build_runner: ^1.12.2
