@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 import 'package:workout_logger/domain/entities/routine_item.dart';
 
 class RoutineDay extends Equatable {
@@ -8,11 +7,12 @@ class RoutineDay extends Equatable {
   final List<RoutineItem> items;
   final int order;
 
-  const RoutineDay(
-      {@required this.id,
-      @required this.name,
-      @required this.items,
-      @required this.order});
+  const RoutineDay({
+    required this.id,
+    required this.name,
+    required this.items,
+    required this.order,
+  });
 
   @override
   List<Object> get props => [name, items, order];
