@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 import 'package:workout_logger/domain/entities/exercise.dart';
 import 'package:workout_logger/domain/entities/set.dart';
 
@@ -12,14 +11,15 @@ class RoutineItem extends Equatable {
   final int order;
   final int groupNumber;
 
-  const RoutineItem(
-      {@required this.id,
-      @required this.exercise,
-      @required this.note,
-      @required this.restDuration,
-      @required this.sets,
-      @required this.order,
-      @required this.groupNumber});
+  const RoutineItem({
+    required this.id,
+    required this.exercise,
+    required this.note,
+    required this.restDuration,
+    required this.sets,
+    required this.order,
+    required this.groupNumber,
+  });
 
   @override
   List<Object> get props =>
